Split Nanoflann and PicoTree benchmarks.
