+++ examples/benchmark/CMakeLists.txt
@@ -1,3 +1,7 @@
+################################################################################
+# bm_nanoflann
+################################################################################
+
 add_executable(bm_nanoflann)
 target_sources(bm_nanoflann
     PRIVATE
@@ -12,11 +16,30 @@ target_link_libraries(bm_nanoflann
 set_target_properties(bm_nanoflann
     PROPERTIES
     RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin
-# For the benchmark demo we want to explicitly use C++11.
-    CXX_STANDARD 11
-    CXX_STANDARD_REQUIRED ON
-    CXX_EXTENSIONS OFF
 )
+
+################################################################################
+# bm_pico_kd_tree
+################################################################################
+
+add_executable(bm_pico_kd_tree)
+target_sources(bm_pico_kd_tree
+    PRIVATE
+        ${CMAKE_CURRENT_LIST_DIR}/bm_pico_kd_tree.cpp
+)
+target_link_libraries(bm_pico_kd_tree
+    PRIVATE
+        pico_common
+        benchmark::benchmark
+)
+set_target_properties(bm_pico_kd_tree
+    PROPERTIES
+    RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin
+)
+
+################################################################################
+# uosr_to_bin
+################################################################################
 
 add_executable(uosr_to_bin)
 target_sources(uosr_to_bin
@@ -32,6 +55,10 @@ set_target_properties(uosr_to_bin
     RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin
 )
 
+################################################################################
+# bin_to_ascii
+################################################################################
+
 add_executable(bin_to_ascii)
 target_sources(bin_to_ascii
     PRIVATE
@@ -45,5 +72,10 @@ set_target_properties(bin_to_ascii
     RUNTIME_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/bin
 )
 
-file(COPY ${CMAKE_CURRENT_LIST_DIR}/plot_benchmarks.py 
-    DESTINATION ${CMAKE_BINARY_DIR}/bin)
+################################################################################
+# plot_benchmarks
+################################################################################
+
+# TODO These don't get deleted when running: make clean
+add_custom_target(plot_benchmarks ALL
+    COMMAND ${CMAKE_COMMAND} -E copy ${CMAKE_CURRENT_LIST_DIR}/plot_benchmarks.py ${CMAKE_BINARY_DIR}/bin)
